Migrate TimeEnum#x to long, cleanup

+++ src/main/java/gg/projecteden/nexus/features/ambience/Wind.java
@@ -17,14 +17,14 @@ public class Wind {
 	@Getter
 	private static double Z;
 
-	private static final int blowingUpdateInterval = TickTime.MINUTE.x(5);
+	private static final long blowingUpdateInterval = TickTime.MINUTE.x(5);
 	private static final Runnable blowingUpdate = () -> {
 		blowing = RandomUtils.chanceOf(25);
 		Ambience.sendDebug("Blowing: " + blowing);
 	};
 
 
-	private static final int directionUpdateInterval = TickTime.MINUTE.get();
+	private static final long directionUpdateInterval = TickTime.MINUTE.get();
 	private static final Runnable directionUpdate = () -> {
 		direction = Math.random() * 2 * Math.PI;
 		X = Math.sin(direction);
