+++ src/main/java/gg/projecteden/nexus/features/events/y2021/bearfair21/fairgrounds/minigolf/MiniGolf.java
@@ -370,7 +370,7 @@ public class MiniGolf {
 							.launch());
 
 						// Send message
-						int wait = TickTime.SECOND.x(2);
+						long wait = TickTime.SECOND.x(2);
 						if (BearFair21.getDailyTokensLeft(user.getPlayer(), BF21PointSource.MINIGOLF, 5) > 0)
 							wait = 0;
